refactor + profile view fix
