+++ Server/src/loaders/loadStartupProcess.ts
@@ -40,7 +40,7 @@ export class loadStartupProcess {
       ],
       credentials: true,
       methods: 'GET,HEAD,OPTIONS,PUT,PATCH,POST,DELETE',
-      origin: "http://localhost:4500",
+      origin: "http://localhost:3000",
       preflightContinue: false,
     };
